+++ components-ui/app/src/tol-ui/src/general/index.ts
@@ -21,6 +21,7 @@ export * from "./IconTooltip";
 export * from "./EditableTitle";
 export * from "./LoadingContent";
 export * from "./Markdown";
+export * from "./NoneValue";
 export * from "./ObjectDetail";
 export * from "./Placeholder";
 export * from "./Plate";