+++ components-ui/app/src/tol-ui/src/interfaces/Filter.ts
@@ -36,7 +36,7 @@ export interface IFilterWithRequiredAnd {
   and_: IAndAttributes;
 }
 
-export type TTranslators = Record<string, string>;
+export type TTranslations = Record<string, string>;
 
 export type TFilterDateRangeValue = [Date, Date] | null;
 
@@ -63,7 +63,16 @@ export type IDBBoardEntityFilter = {
    */
   filter_exclude_incoming?: boolean;
   /**
+   * Whether relationship translations are enabled for the zone.
+   */
+  relationship_translation?: boolean;
+  /**  /**
+   * Custom relationship translation path for incoming filters, used at the zone level.
+   * This is a single path string (e.g., "specimen.species") that can be used to translate.
+   */
+  translation_path?: string;
+  /**
    * Custom translations for incoming filters, used at the zone level.
    */
-  translations?: TTranslators;
+  attribute_translations?: TTranslations;
 };