+++ components-ui/app/src/tol-ui/src/table/cellRenderers/CellRendererConditionParamOptions.tsx
@@ -11,7 +11,7 @@ import {
   BUTTONS,
   cellRendererParams,
   deepCopy,
-  generateDefaultFilter,
+  createEmptyFilter,
   defineZoneWithComponentList,
   generateFilter,
   isEmptyObject,
@@ -74,7 +74,7 @@ export function CellRendererConditionParamOptions(props: PCellRendererConditionP
   const zoneFilterId = "cell-renderer-component";
   const [filterZone, setFilterZone] = useState<IZone>(
     defineZoneWithComponentList("dummy-object-for-remote-filters", [
-      { id: zoneFilterId, filter: renderer?.props?.[paramName!] as IFilter || generateDefaultFilter() },
+      { id: zoneFilterId, filter: renderer?.props?.[paramName!] as IFilter || createEmptyFilter() },
     ]),
   );
   // The filters used for this condition
@@ -92,7 +92,7 @@ export function CellRendererConditionParamOptions(props: PCellRendererConditionP
   useEffect(() => {
     if (renderer?.props?.[paramName!] && paramName) {
       const paramValue = renderer.props[paramName];
-      const filterValue = typeof paramValue === 'object' ? paramValue as IFilter : generateDefaultFilter();
+      const filterValue = typeof paramValue === 'object' ? paramValue as IFilter : createEmptyFilter();
       setAttributes(Object.keys(filterValue.and_ || {}) || []);
       setFilterConditions(filterValue);
       setPreviousFilterConditions(previousRenderer?.props?.[paramName] as IFilter | undefined);
@@ -101,8 +101,8 @@ export function CellRendererConditionParamOptions(props: PCellRendererConditionP
       ]));
     } else {
       setAttributes([]);
-      setFilterConditions(generateDefaultFilter());
-      setPreviousFilterConditions(generateDefaultFilter());
+      setFilterConditions(createEmptyFilter());
+      setPreviousFilterConditions(createEmptyFilter());
     }
   }, [paramName]);
 
@@ -120,7 +120,7 @@ export function CellRendererConditionParamOptions(props: PCellRendererConditionP
   // Needed for the BottomButtons. Either adding the parameter or going back means that the filter
   // making up the parameter should be cleared (it gets tied to the renderer somehow!)
   const resetFilterZone = () => setFilterZone(defineZoneWithComponentList("dummy-object-for-remote-filters", [
-    { id: zoneFilterId, filter: generateDefaultFilter() },
+    { id: zoneFilterId, filter: createEmptyFilter() },
   ]));
 
   const handleBack = () => {